Testkit utility for creating and dropping database
I had planned to just document how to do this, and use the emutlator,
using the actual spanner client.

Unfortunately the spanner client can't be on the same classpath as APS
as it has a conficting version of grpc java.

Instead offering a very basic way to create and drop the schema via
our gRPC client.