Added object changes support
This adds changes and currentChanges queries to the object store.

This change doesn't change any existing storage structure or format,
except that it does add an index - however, given that
`akka-persistence-spanner` doesn't actually create any tables/indexes
for you, that's up to consumers of the library, the creation of that
index is opt-in. The index does create a hotspot on inserting/updating
objects for the same entity type.